dev-requirements.txt
flake8
pytest-ckan
pytest-factoryboy
